Make `named()` require explicit extension
The `named` method seems to have been designed to be extension-agnostic,
but the only place it's used is in `ClassSearch`, which passes "Foo.php"
to find matching PHP classes. The unintended consequence of that was it
also got results with `.php.html` extensions, so this removes the `.*`.

Fixes phpactor/phpactor#1216